+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Added
 
+- Added hook ENTITY:ClientUse(), which is triggered clientside if an entity is used
+  - Return true to prevent also using this on the server for clientside only usecases
+
 ### Changed
 
 ### Fixed